refactor: cutoff base responses
