+++ frontend/src/types/base.ts
@@ -1,0 +1,8 @@
+export type BaseResponse = {
+  success: true
+}
+
+export type ErrorResponse =  {
+  success: false
+  detail: string
+}
